+++ src/modules/user/page/login/interfaces.ts
@@ -1,0 +1,8 @@
+import { OAuthType } from 'Models/oauth/type';
+import { FormLogin, FormLoginError } from './types';
+
+export interface Props {
+  onSubmit?: (form: FormLogin) => void;
+  onError?: (form: FormLoginError) => void;
+  onSubmitOAuth?: (type: OAuthType) => void;
+}
