+++ src/modules/user/page/login/types.ts
@@ -1,0 +1,2 @@
+export type FormLogin = { email: string; password: string };
+export type FormLoginError = { values: FormLogin; errorFields: [] };
